# finance-manager-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API serves personal finance endpoints. Accuracy is not garunteed",
        "title": "Go Finance Manager API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/users/{userId}/roles": {
            "get": {
                "description": "Returns an array of UserRole objects belonging to a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Roles"
                ],
                "summary": "Get All User Roles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user we are searching for",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search for user role by role name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserRole"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/roles/{roleId}": {
            "post": {
                "description": "Adds a new role to a User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Roles"
                ],
                "summary": "Add User Role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user to add a role to",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the role to add to the user",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes a role from a a User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Roles"
                ],
                "summary": "Remove User Role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user to remove a role from",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the role to remove from the user",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jsonutils.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.UserRole": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "roleId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}